Removed Brocfile in favor of new ember-cli-build
