--- Brocfile.js
@@ -1,36 +1,0 @@
-/* global require, module */
-
-var EmberApp = require('ember-cli/lib/broccoli/ember-app');
-var pickFiles = require('broccoli-static-compiler');
-
-var app = new EmberApp();
-
-// Use `app.import` to add additional libraries to the generated
-// output files.
-//
-// If you need to use different assets in different
-// environments, specify an object as the first parameter. That
-// object's keys should be the environment name and the values
-// should be the asset to use in that environment.
-//
-// If the library that you are including contains AMD or ES6
-// modules that you would like to import into your application
-// please specify an object with the list of modules as keys
-// along with the exports of each module as its value.
-
-// Foundation
-app.import('bower_components/foundation/js/foundation.min.js');
-
-// jQuery Sortable
-app.import('bower_components/jquery-ui/ui/core.js');
-app.import('bower_components/jquery-ui/ui/widget.js');
-app.import('bower_components/jquery-ui/ui/mouse.js');
-app.import('bower_components/jquery-ui/ui/sortable.js');
-
-var fontawesomeAssets = pickFiles('bower_components/fontawesome', {
-    srcDir: '/fonts',
-    files: ['FontAwesome.otf', 'fontawesome-webfont.*'],
-    destDir: 'assets/fontawesome/fonts'
-});
-
-module.exports = app.toTree(fontawesomeAssets);
